Refactor pressableOwnership into a custom hook
